dm3058-capture: Remove ping; instead handle initial HTTP connection error
Doing a ping before we connect is rather pointless since our initial
HTTP connection will also generate a similar error if the host is not
online. So drop the ping and instead handle the HTTP connection error.

This both reduces complexity and saves running an external program in
a subprocess, increasing reliability.

The current code for handling this is not optimal, but the general
error handling of this needs some major reworking. For example,
currently (as before) attempting to connect to a host that's up but
doesn't have a web server running, or that does but is not a DM3058,
will simply print "Capture failed" rather than something more
informative (such as, "That's a DS1054Z, silly!").

One might also consider that there may be several Rigol hosts on the
network (say, both a DZ1054Z and a DM3058). If not given an address
currently the code that searches for Rigol devices on the network
might pick the wrong one; it would be nice if it could instead query
each device in turn until it finds the right one.