{ds1054,e36312a}-capture: Remove ping and handle TCP connection error
There's no need to ping the host before trying to connect via TCP
since the TCP connection itself gives us all the error information we
need. So drop the ping (which removes it from common.py as these were
the last users of it) and just handle connection errors from the TCP
connnect attempt.

This also now gives just an error message, not an exception traceback,
when the host is up but is not listening on the port on which we wish
to connect.