WIP *-capture: Extract common filename generation code
This code still generates filenames with colons in them, which will
break on Windows. (This wasn't previously an issue when we were
dependent on being able to run Linux programs such as `arp` and
`grep`, but those dependencies are disappearing, giving this the
potential to be multi-platform.)

This also slightly changes the code used by ds1054-capture; it now
does not prepend os.getcwd() to the filename it generates. That should
make no difference, however.